Add named commands and specific functions
add constants for possible server side operations.
add functions for doing specific operations in
the server side, for example activating a phone
or loading the list of files to put in the sdcard.
this way we have all server operations centralized,
and other classes need no knowledge of server
commands or parameters. response is always a json
object.